requirements.txt
Django==5.1
djangorestframework==3.15.2

django-cors-headers==4.5.0
dj-database-url==2.2.0
gunicorn==23.0.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
whitenoise==6.7.0
